pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[project]
name = "deva"
version = "1.0.0"
authors = [{ name = "Rex Cheng", email = "hkchengrex@gmail.com" }]
description = "Tracking Anything with Decoupled Video Segmenttation (DEVA), ICCV 2023"
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
  "Programming Language :: Python :: 3",
  "Operating System :: OS Independent",
]
dependencies = [
  'gitpython',
  'thinplate@git+https://github.com/cheind/py-thin-plate-spline',
  'hickle',
  'tensorboard',
  'numpy',
  'Pillow',
  'opencv-python',
  'scipy',
  'pycocotools',
  'supervision',
  'tqdm',
  'gurobipy',
  'PuLP',
  'gradio',
  'gdown',
]

[project.urls]
"Homepage" = "https://github.com/hkchengrex/Tracking-Anything-with-DEVA"
"Bug Tracker" = "https://github.com/hkchengrex/Tracking-Anything-with-DEVA/issues"
